Return float64 weights from trapezoidal_weights (#243)
torch.ones(n) picks up the default dtype, so the weights came back float32
while the nodes, built from np.linspace, were float64. This was the only rule
in the quadrature module with that mismatch: legendre_gauss_weights,
lobatto_weights and clenshaw_curtiss_weights are all float64 throughout.

It capped the accuracy of everything derived from the rule at around 1e-7.
The visible public path is precompute_latitudes(grid="equiangular-trapezoidal"),
whose weights were float32 while the other three grids were float64. It is also
inherited by any rule built on top of the trapezoidal one, where the mismatch
can hide: multiplying float64 nodes by float32 weights promotes back to float64,
so the result carries float32 rounding behind a float64 dtype.

Measured on QuadratureS2 integrating f = 1 over the sphere in float64, the
equiangular-trapezoidal error drops from 8.4e-07 to 9.7e-08, in line with the
8.2e-08 of the equiangular grid. The remaining error is the layer casting its
quad_weight buffer to float32, which this change does not touch.

Tests: the two equiangular-trapezoidal cases in test_constant_integral no
longer need their loosened 1e-4 tolerance and now use the 1e-6 of the other
grids, since the measured error is identical to equiangular's. Note the
tightening alone does not catch the bug, as the rtol term leaves enough slack
either way; the new TestQuadratureWeightPrecision cases do, comparing the
weights against a float64 analytic reference at 1e-15 and asserting dtype
consistency across all four grids. They fail on main and pass here.

test_polynomial_latitude_integral keeps its 1e-2 trapezoidal tolerance: that
error is the genuine O(h^2) of the rule, measured at 2.1e-03, not a precision
artifact.

+++ torch_harmonics/quadrature.py
@@ -162,7 +162,9 @@ def trapezoidal_weights(n: int, a: Optional[float] = -1.0, b: Optional[float] = 
     """
 
     xlg = torch.as_tensor(np.linspace(a, b, n, endpoint=not periodic))
-    wlg = (b - a) / (n - 1 + periodic * 1) * torch.ones(n, requires_grad=False)
+    # dtype is explicit: torch.ones defaults to float32, which would make these the only
+    # float32 weights in the module and cap the accuracy of every rule derived from them
+    wlg = (b - a) / (n - 1 + periodic * 1) * torch.ones(n, dtype=xlg.dtype, requires_grad=False)
 
     if not periodic:
         wlg[0] *= 0.5
